rename content manager parts
