+++ src/web/views/Content/List/style.scss
@@ -1,61 +1,96 @@
 .list {
   min-height: 85vh;
-  // background-color: $content-bg-color;
-  padding-right: 60px;
-  padding-left: 6px;
 
-  button {
-    margin-left: 6px;
-    margin-top: 18px;
-    width: 100%;
+  .header {
+    padding: 10px 0;
+    height: 60px;
+    margin: 5px 0 0 0;
+    border-bottom: 1px solid #c9c8ca;
 
-    &.disabled {
-      background-color: #b5b3b6;
-      cursor: not-allowed;
+    :global(.checkbox) {
+      margin-top: 0px;
+      margin-bottom: 0px;
+      margin-left: -2px;
+      height: 18px;
+    }
 
-      &:hover {
-        background-color: #b5b3b6;
+    button {
+      height: 30px;
+      width: 34px;
+      position: relative;
+
+      i {
+        position: absolute;
+        left: 6px;
+        top: 6px;
+        font-size: 18px;
+        line-height: 18px;
+
+        color: #807f81;
+      }
+
+      margin-right: 8px;
+    }
+
+    .left {
+      .search {
+        position: relative;
+        width: 300px;
+        display: inline-block;
+        margin-right: 10px;
+        margin-left: 10px;
+
+        :global(.form-control) {
+          position: absolute;
+          top: -4px;
+
+          height: 30px;
+        }
+      }
+    }
+
+    .right {
+      right: 5px;
+
+      .paging {
+        margin-right: 10px;
       }
     }
   }
 
-  ul {
-    list-style: none;
-    padding: 10px;
-    margin-top: 20px;
+  .container {
+    table {
+      border: none;
 
-    li {
-      margin-bottom: 5px;
-      padding-left: 6px;
-      line-height: 24px;
-      border-left: solid 3px #fff;
+      tr {
+        font-size: 90%;
+      }
 
-      a {
+      td {
+        padding: 10px;
+        border: none;
+        border-bottom: 1px solid #e0e0e0;
+      }
+
+      :global(.checkbox) {
+        margin-top: 0px;
+        margin-bottom: 0px;
+        height: 20px;
+      }
+
+      .item {
         cursor: pointer;
-        color: #000000;
-        font-weight: bolder;
-        text-decoration: none;
-
-        &:hover {
-          color: #444444;
-        }
       }
 
-      &.selected {
-        border-color: #46cba0;
-
-        a {
-          font-weight: bold;
-        }
-      }
-
-      .title {
-        margin-right: 5px;
-      }
-
-      .count {
-        opacity: 0.7;
+      .selected {
+        background-color: rgb(125, 228, 194);
       }
     }
+  }
+
+  .empty {
+    text-align: center;
+    font-weight: bolder;
+    margin-top: 35px;
   }
 }
