--- src/web/views/Content/Manage/style.scss
@@ -1,97 +1,0 @@
-.manage {
-  min-height: 85vh;
-  // background-color: $content-bg-color;
-
-  .header {
-    padding: 10px 0;
-    height: 60px;
-    margin: 5px 0 0 0;
-    border-bottom: 1px solid #c9c8ca;
-
-    :global(.checkbox) {
-      margin-top: 0px;
-      margin-bottom: 0px;
-      margin-left: -2px;
-      height: 18px;
-    }
-
-    button {
-      height: 30px;
-      width: 34px;
-      position: relative;
-
-      i {
-        position: absolute;
-        left: 6px;
-        top: 6px;
-        font-size: 18px;
-        line-height: 18px;
-
-        color: #807f81;
-      }
-
-      margin-right: 8px;
-    }
-
-    .left {
-      .search {
-        position: relative;
-        width: 300px;
-        display: inline-block;
-        margin-right: 10px;
-        margin-left: 10px;
-
-        :global(.form-control) {
-          position: absolute;
-          top: -4px;
-
-          height: 30px;
-        }
-      }
-    }
-
-    .right {
-      right: 5px;
-
-      .paging {
-        margin-right: 10px;
-      }
-    }
-  }
-
-  .container {
-    table {
-      border: none;
-
-      tr {
-        font-size: 90%;
-      }
-
-      td {
-        padding: 10px;
-        border: none;
-        border-bottom: 1px solid #e0e0e0;
-      }
-
-      :global(.checkbox) {
-        margin-top: 0px;
-        margin-bottom: 0px;
-        height: 20px;
-      }
-
-      .item {
-        cursor: pointer;
-      }
-
-      .selected {
-        background-color: rgb(125, 228, 194);
-      }
-    }
-  }
-
-  .empty {
-    text-align: center;
-    font-weight: bolder;
-    margin-top: 35px;
-  }
-}
