+++ src/web/views/Content/Sidebar/style.scss
@@ -1,0 +1,60 @@
+.sidebar {
+  min-height: 85vh;
+  padding-right: 60px;
+  padding-left: 6px;
+
+  button {
+    margin-left: 6px;
+    margin-top: 18px;
+    width: 100%;
+
+    &.disabled {
+      background-color: #b5b3b6;
+      cursor: not-allowed;
+
+      &:hover {
+        background-color: #b5b3b6;
+      }
+    }
+  }
+
+  ul {
+    list-style: none;
+    padding: 10px;
+    margin-top: 20px;
+
+    li {
+      margin-bottom: 5px;
+      padding-left: 6px;
+      line-height: 24px;
+      border-left: solid 3px #fff;
+
+      a {
+        cursor: pointer;
+        color: #000000;
+        font-weight: bolder;
+        text-decoration: none;
+
+        &:hover {
+          color: #444444;
+        }
+      }
+
+      &.selected {
+        border-color: #46cba0;
+
+        a {
+          font-weight: bold;
+        }
+      }
+
+      .title {
+        margin-right: 5px;
+      }
+
+      .count {
+        opacity: 0.7;
+      }
+    }
+  }
+}
